Show archived agenda items and deliverables by url
When you go to the specific url of an archived agenda item
or deliverable you no longer get a broken page.
You can't see the timeline contents at the moment but at least
the page is no longer broken.
That was also a problem when you clicked on an archived agenda
item or deliverable from within a timeline.