Removed archived_at for conv and org member tables
Archiving (and not deleting) conv and org member objects
triggered a side effect of no be allowed to add the same user
to a conv or org again. This is due to a unique constraint in the
database. To support this in a good way it would take more hours
so I report that upstream and let them decide about the priority.
I think it makes sense to implement that in a archived and good way
when we implement the recover feature for objects.