Moved post-db work for comments to command
This is the first usage of the new command pattern that we
are using for handline user interactions.
It gives us one place to handle the interactions after
the object itself got persisted.
It also cleans up the controller tier a lot.

With this commit we also added a new link table called
TimelineUser which holds information like the subscription level.
In the future it might also hold the last-seen marker and
other metadata that is used in the stream view.